news_chatbot/voice_backend/pyproject.toml
[project]
name = "voice-backend"
version = "0.1.0"
description = "News chatbot voice backend with pipecat multi-transport support"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[daily,deepgram,cartesia,elevenlabs,openai,silero,websocket]>=0.0.98",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.11.0",
    "aiortc>=1.10.0",
    "opencv-python>=4.8.0",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
